requirements.txt
python-barcode
fire
pyzbar
